requirements.txt
Django==5.2.10
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.0
django-filter==25.2
drf-spectacular==0.29.0
drf-nested-routers==0.95.0

celery==5.6.2
django-celery-beat==2.8.1
redis==7.1.0
django-redis==6.0.0
gunicorn==21.2.0

psycopg2==2.9.11

python-dotenv==1.2.1
PyYAML==6.0.3
sqlparse==0.5.5
pytz==2025.2
tzdata==2025.3
tzlocal==5.3.1

uritemplate==4.2.0
inflection==0.5.1

flake8==7.1.1
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
black==25.12.0
coverage==7.13.1

pyTelegramBotAPI==4.29.1

stripe==14.1.0
